nephele_SQlite_testing/vo-wot/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "vo-wot"
description = "Virtual Object python implementation using a W3C WoT Runtime and the WoT Scripting API"
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
  { name = 'Andres Garcia Mangas', email = 'andres.garcia@fundacionctic.org' }
]
maintainers = [
  { name = "Nikos Filinis", email = "nfilinis@netmode.ntua.gr" },
  { name = "Dimitris Spatharakis", email = "dspatharakis@netmode.ntua.gr" }
]
keywords = ["wot", "iot", "gateway", "fog", "w3c", "vo", "virtual object"]

classifiers = [
  'Development Status :: 4 - Beta',
  'Intended Audience :: Developers',
  'Natural Language :: English',
  'Operating System :: OS Independent',
  'License :: OSI Approved :: MIT License',
  'Programming Language :: Python :: 3.8',
  'Programming Language :: Python :: 3.9',
  'Programming Language :: Python :: 3.10',
  'Programming Language :: Python :: 3.11'
]
dependencies = [
  'tornado>=6.2,<7.0',
  'jsonschema>=4.17.3,<5.0',
  'reactivex>=4.0.4,<5.0',
  'python-slugify>=8.0.0,<9.0',
  'requests-oauthlib>=1.3.1,<1.4',
  'influxdb-client>=1.36.1,<1.37',
  'pmdarima==2.0.4',
  'numpy==1.26.4',
  'pyOpenSSL>=23.0.0,<24.0.0',
  'aiocoap[linkheader,oscore]==0.4.7 ; platform_system != "Darwin"',
  'amqtt==0.11.0b1',
  'websockets>=8.0'
]

[project.optional-dependencies]
tests = [
  'pytest>=7.2.1',
  'pytest-cov>=4.0.0,<5.0.0',
  'pytest-rerunfailures>=11.1.1,<12.0',
  'faker>=17.0.0,<18.0.0',
  'coveralls>=3.3.1,<4.0',
  'coverage>=6.5.0,<7.0',
  'autopep8>=2.0.1,<3.0',
  'rope>=1.7.0,<2.0'
]

docs = [
  'Sphinx>=6.1.3,<7.0.0',
  'sphinx-rtd-theme>=1.2.0,<2.0.0'
]

[project.scripts]
vo-wot = "wotpy.cli.cli:main"

[project.urls]
Documentation = "https://netmode.gitlab.io/vo-wot/"
Repository = "https://gitlab.eclipse.org/eclipse-research-labs/nephele-project/vo-wot.git"

[tool.setuptools.dynamic]
version = {attr = "wotpy.__version__.__version__"}

[tool.setuptools.packages.find]
where = [ "." ]
include = ["wotpy", "wotpy.*"]
